+++ blocktracker/blocktracker_test.go
@@ -2,8 +2,6 @@ package blocktracker
 
 import (
 	"context"
-	"log"
-	"os"
 	"testing"
 	"time"
 
@@ -53,27 +51,27 @@ func testListener(t *testing.T, server *testutil.TestServer, tracker BlockTracke
 }
 
 func TestBlockTracker_Listener_JsonRPC(t *testing.T) {
-	t.Skip()
+	t.Skip("Too brittle on CI, FIX")
+
 	s := testutil.NewTestServer(t)
 
 	c, _ := jsonrpc.NewClient(s.HTTPAddr())
 	defer c.Close()
 
-	logger := log.New(os.Stderr, "", log.LstdFlags)
-	tracker := NewJSONBlockTracker(logger, c.Eth())
+	tracker := NewJSONBlockTracker(c.Eth())
 	tracker.PollInterval = 1 * time.Second
 	testListener(t, s, tracker)
 }
 
 func TestBlockTracker_Listener_Websocket(t *testing.T) {
-	t.Skip()
+	t.Skip("Too brittle on CI, FIX")
+
 	s := testutil.NewTestServer(t)
 
 	c, _ := jsonrpc.NewClient(s.WSAddr())
 	defer c.Close()
 
-	logger := log.New(os.Stderr, "", log.LstdFlags)
-	tracker, err := NewSubscriptionBlockTracker(logger, c)
+	tracker, err := NewSubscriptionBlockTracker(c)
 	if err != nil {
 		t.Fatal(err)
 	}