+++ jsonrpc/eth_test.go
@@ -10,6 +10,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/umbracle/ethgo"
+	"github.com/umbracle/ethgo/abi"
 	"github.com/umbracle/ethgo/testutil"
 )
 
@@ -245,6 +246,54 @@ func TestEthChainID(t *testing.T) {
 	})
 }
 
+func TestEthCall(t *testing.T) {
+	s := testutil.NewTestServer(t)
+
+	c, _ := NewClient(s.HTTPAddr())
+	cc := &testutil.Contract{}
+
+	// add global variables
+	cc.AddCallback(func() string {
+		return "uint256 val = 1;"
+	})
+
+	// add setter method
+	cc.AddCallback(func() string {
+		return `function getValue() public returns (uint256) {
+			return val;
+		}`
+	})
+
+	_, addr, err := s.DeployContract(cc)
+	require.NoError(t, err)
+
+	input := abi.MustNewMethod("function getValue() public returns (uint256)").ID()
+
+	resp, err := c.Eth().Call(&ethgo.CallMsg{To: &addr, Data: input}, ethgo.Latest)
+	require.NoError(t, err)
+
+	require.Equal(t, "0x0000000000000000000000000000000000000000000000000000000000000001", resp)
+
+	nonce := uint64(1)
+
+	// override the state
+	override := &ethgo.StateOverride{
+		addr: ethgo.OverrideAccount{
+			Nonce:   &nonce,
+			Balance: big.NewInt(1),
+			StateDiff: &map[ethgo.Hash]ethgo.Hash{
+				// storage slot 0 stores the 'val' uint256 value
+				{0x0}: {0x3},
+			},
+		},
+	}
+
+	resp, err = c.Eth().Call(&ethgo.CallMsg{To: &addr, Data: input}, ethgo.Latest, override)
+	require.NoError(t, err)
+
+	require.Equal(t, "0x0300000000000000000000000000000000000000000000000000000000000000", resp)
+}
+
 func TestEthGetNonce(t *testing.T) {
 	s := testutil.NewTestServer(t)
 
@@ -345,10 +394,46 @@ func TestEthFeeHistory(t *testing.T) {
 	lastBlock, err := c.Eth().BlockNumber()
 	assert.NoError(t, err)
 
-	from := ethgo.BlockNumber(lastBlock - 2)
-	to := ethgo.BlockNumber(lastBlock)
-
-	fee, err := c.Eth().FeeHistory(from, to)
+	fee, err := c.Eth().FeeHistory(1, ethgo.BlockNumber(lastBlock), []float64{25, 75})
 	assert.NoError(t, err)
 	assert.NotNil(t, fee)
+}
+
+func TestEthMaxPriorityFeePerGas(t *testing.T) {
+	s := testutil.NewTestServer(t)
+	c, err := NewClient(s.HTTPAddr())
+	require.NoError(t, err)
+
+	initialMaxPriorityFee, err := c.Eth().MaxPriorityFeePerGas()
+	require.NoError(t, err)
+
+	// wait for 2 blocks
+	require.NoError(t, s.ProcessBlock())
+	require.NoError(t, s.ProcessBlock())
+
+	txn := &ethgo.Transaction{
+		To:                   &testutil.DummyAddr,
+		Value:                ethgo.Gwei(1),
+		Type:                 ethgo.TransactionDynamicFee,
+		MaxPriorityFeePerGas: ethgo.Gwei(1),
+	}
+
+	latestBlock, err := c.Eth().BlockNumber()
+	require.NoError(t, err)
+
+	feeHistory, err := c.Eth().FeeHistory(1, ethgo.BlockNumber(latestBlock), nil)
+	require.NoError(t, err)
+
+	latestBaseFee := feeHistory.BaseFee[len(feeHistory.BaseFee)-1]
+	txn.MaxFeePerGas = new(big.Int).Add(latestBaseFee, txn.MaxPriorityFeePerGas)
+
+	receipt, err := s.SendTxn(txn)
+	require.NoError(t, err)
+	require.Equal(t, uint64(1), receipt.Status)
+
+	newMaxPriorityFee, err := c.Eth().MaxPriorityFeePerGas()
+	t.Log(initialMaxPriorityFee)
+	t.Log(newMaxPriorityFee)
+	require.NoError(t, err)
+	require.True(t, initialMaxPriorityFee.Cmp(newMaxPriorityFee) <= 0)
 }