+++ wallet/signer_test.go
@@ -5,8 +5,53 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 	"github.com/umbracle/ethgo"
+	"pgregory.net/rapid"
 )
+
+func TestSigner_SignAndRecover(t *testing.T) {
+	rapid.Check(t, func(t *rapid.T) {
+		// fill in common types for a transaction
+		txn := &ethgo.Transaction{}
+
+		if rapid.Bool().Draw(t, "to") {
+			to := ethgo.BytesToAddress(rapid.SliceOf(rapid.Byte()).Draw(t, "to_addr"))
+			txn.To = &to
+		}
+
+		txType := rapid.IntRange(0, 2).Draw(t, "tx type")
+
+		// fill in specific fields depending on the type
+		// of the transaction.
+		txn.Type = ethgo.TransactionType(txType)
+		if txn.Type == ethgo.TransactionDynamicFee {
+			maxFeePerGas := rapid.Int64Range(1, 1000000000).Draw(t, "maxFeePerGas")
+			txn.MaxFeePerGas = big.NewInt(maxFeePerGas)
+			maxPriorityFeePerGas := rapid.Int64Range(1, 1000000000).Draw(t, "maxPriorityFeePerGas")
+			txn.MaxPriorityFeePerGas = big.NewInt(maxPriorityFeePerGas)
+		} else {
+			gasPrice := rapid.Uint64Range(1, 1000000000).Draw(t, "gasPrice")
+			txn.GasPrice = gasPrice
+		}
+
+		// signer is from a random chain
+		chainId := rapid.Uint64().Draw(t, "chainId")
+		signer := NewEIP155Signer(chainId)
+
+		key, err := GenerateKey()
+		require.NoError(t, err)
+
+		signedTxn, err := signer.SignTx(txn, key)
+		require.NoError(t, err)
+
+		// recover the sender
+		sender, err := signer.RecoverSender(signedTxn)
+		require.NoError(t, err)
+
+		require.Equal(t, sender, key.Address())
+	})
+}
 
 func TestSigner_EIP1155(t *testing.T) {
 	signer1 := NewEIP155Signer(1337)
@@ -26,14 +71,6 @@ func TestSigner_EIP1155(t *testing.T) {
 	from, err := signer1.RecoverSender(txn)
 	assert.NoError(t, err)
 	assert.Equal(t, from, key.addr)
-
-	/*
-		// try to use a signer with another chain id
-		signer2 := NewEIP155Signer(2)
-		from2, err := signer2.RecoverSender(txn)
-		assert.NoError(t, err)
-		assert.NotEqual(t, from, from2)
-	*/
 }
 
 func TestTrimBytesZeros(t *testing.T) {